+++ src/e_checks.cc
@@ -1424,7 +1424,7 @@ void Things_FindInVoid(selection_c& list)
 
 		Objid obj;
 
-		GetNearObject(obj, OBJ_SECTORS, x, y);
+		GetNearObject(obj, OBJ_SECTORS, static_cast<float>(x), static_cast<float>(y));
 
 		if (! obj.is_nil())
 			continue;
@@ -1443,7 +1443,7 @@ void Things_FindInVoid(selection_c& list)
 			int x2 = x + ((corner & 1) ? -4 : +4);
 			int y2 = y + ((corner & 2) ? -4 : +4);
 
-			GetNearObject(obj, OBJ_SECTORS, x2, y2);
+			GetNearObject(obj, OBJ_SECTORS, static_cast<float>(x2), static_cast<float>(y2));
 
 			if (obj.is_nil())
 				out_count++;
