+++ src/im_img.cc
@@ -805,18 +805,18 @@ Img_c * IM_CreateLightSprite()
 			float dx = (W - 2*x) / (float)W;
 			float dy = (H - 2*y) / (float)H;
 
-			float dist = sqrt((dx) * (dx) + (dy) * (dy));
+			float dist = sqrtf((dx) * (dx) + (dy) * (dy));
 
-			float ity = 1.0 / (dist + 0.5) / (dist + 0.5);
+			float ity = 1.0f / (dist + 0.5f) / (dist + 0.5f);
 
 			if (ity < 0.5)
 				continue;
 
-			ity = (ity - 0.4) / (1.0 - 0.4);
+			ity = (ity - 0.4f) / (1.0f - 0.4f);
 
-			int r = 255 * ity;
-			int g = 235 * ity;
-			int b = 90  * ity;
+			int r = static_cast<int>(255 * ity);
+			int g = static_cast<int>(235 * ity);
+			int b = static_cast<int>(90  * ity);
 
 			pix = W_FindPaletteColor(r, g, b);
 		}
